Swap to using Webpack instead of Brunch
This is so we can use [Elm Webpack Loader][1], which allows us to `npm install elm` and then use the binary that npm installed.

With brunch, we require Elm to be installed on the OS. This is not possible if deployment is done via Dokku (and if it is, I've wasted too much time trying to do it).

[1]: https://github.com/rtfeldman/elm-webpack-loader